Added _generator.sh script
Added a bash script which parses the output of the `keybase help` and `keybase help command` to retrieve a list of subcommands. It then uses this output to generate a precomputed, static, bash-completion script based on the one you have here. This would just make it easier to automate rebuilding the completion whenever the commands change.

If that's outside of the scope of your project, that's fine, just thought it might be useful.